+++ Makefile
@@ -1,0 +1,2 @@
+all:
+	gcc startmenu.c maingame.c hardcore.c filereader.c boardfunctions.c tutorialpages.c -o sudoku.exe
