+++ headers/tutorialpages.h
@@ -1,0 +1,12 @@
+#ifndef __TUTORIAL_PAGES__
+#define __TUTORIAL_PAGES__
+
+#include "boardfunctions.h"
+
+typedef void (*PageFn)(int*);
+
+void page1(int *pageno);
+void page2(int *pageno);
+void page3(int *pageno);
+
+#endif
